Fix issue #4
If you refresh the page that has "page=N" (where N is some number), then Batchiepatchie will load it as a string.

When you press next page it'll end up doing stuff like "page + 1" which will do thing like "page=3" turn into "page=31" instead of "page=4".

This commit makes it so that the page is parsed as integer when loaded from query parameters.