Fix very high memory use when looking at very large logs.
We found an issue when looking at a job with very large logs from the UI.

Batchiepatchie backend go (in handlers.go FetchLogs) will fetch the log entirely
to memory, and then send it as JSON to the browser.

If the logs are large enough, this can be massive memory sink. We couldn't run
batchiepatchie reliably on 40 gigabytes of memory.

This commit makes 'text' type logs output the only supported format, because it
can be streamed easily (it's just newline-delimited messages).

In the backend we removed support for JSON logs and in frontend we started asking
for text-type logs in all queries.