Fix frontend Dockerfile from not building.
It broke because I merged incompatible changes related to PR #6. PR #6
made a fix to the Dockerfile related to Debian distribution Jessie not
being hosted on debian mirror network.

The fix made by PR #6 is no longer elevant because we updated the base
Node image. So this PR removes the hack.