Fix compilation error by making logentries host configurable.
A dependency has updated that changed how logentries logger is
initialized. You now need to provide a hostname when you instantiate
this logger.

This commit knows the default logentries host, and lets you override it
in configuration file if needed for whatever reason.

The new config item is called "logentries_host".